scraper/requirements.txt
selenium>=4.15.0
requests>=2.31.0
Pillow>=10.1.0
pytesseract>=0.3.10
google-generativeai>=0.3.0
python-dotenv>=1.0.0
